Fix/nafi/inter service comm

+++ SearchService/src/main/java/com/ynm/searchservice/Repository/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import java.util.List;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     // find by email
     Optional<User> findByEmail(String email);
